refactor(context): split into modules
